updated the eval command for more security And deleted duplicated
